[Generator] Clean the attribute work.
Some changes:

1. Move all the static attr related methods to the attr factory.
2. Make it simpler to see the diff between NET and not by using a
   partial class.
